Revert "initial buggy thing"
This reverts commit 3b1c37f6aea8b1be6275043936838ef0f704a133.
