+++ src/OpenXcom.2010.vcxproj
@@ -390,7 +390,6 @@
     <ClCompile Include="FTA\MasterMind.cpp" />
     <ClCompile Include="Geoscape\AlienBaseState.cpp" />
     <ClCompile Include="Geoscape\AllocateTrainingState.cpp" />
-    <ClCompile Include="Geoscape\AltMonthlyReportState.cpp" />
     <ClCompile Include="Geoscape\CraftNotEnoughPilotsState.cpp" />
     <ClCompile Include="Geoscape\DogfightErrorState.cpp" />
     <ClCompile Include="Geoscape\DogfightExperienceState.cpp" />
@@ -799,7 +798,6 @@
     <ClInclude Include="FTA\MasterMind.h" />
     <ClInclude Include="Geoscape\AlienBaseState.h" />
     <ClInclude Include="Geoscape\AllocateTrainingState.h" />
-    <ClInclude Include="Geoscape\AltMonthlyReportState.h" />
     <ClInclude Include="Geoscape\Cord.h" />
     <ClInclude Include="Geoscape\CraftNotEnoughPilotsState.h" />
     <ClInclude Include="Geoscape\DogfightErrorState.h" />
