+++ src/OpenXcom.2010.vcxproj.filters
@@ -624,6 +624,7 @@
     <ClCompile Include="Interface\ToggleTextButton.cpp">
       <Filter>Interface</Filter>
     </ClCompile>
+    <ClCompile Include="pch.cpp" />
     <ClCompile Include="Savegame\SerializationHelper.cpp">
       <Filter>Savegame</Filter>
     </ClCompile>
@@ -1177,9 +1178,6 @@
     </ClCompile>
     <ClCompile Include="Basescape\SoldierTransformationListState.cpp">
       <Filter>Basescape</Filter>
-    </ClCompile>
-    <ClCompile Include="Geoscape\AltMonthlyReportState.cpp">
-      <Filter>Geoscape</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
@@ -1815,6 +1813,7 @@
     <ClInclude Include="Interface\ToggleTextButton.h">
       <Filter>Interface</Filter>
     </ClInclude>
+    <ClInclude Include="pch.h" />
     <ClInclude Include="Engine\ShaderDraw.h">
       <Filter>Engine</Filter>
     </ClInclude>
@@ -2422,9 +2421,6 @@
     </ClInclude>
     <ClInclude Include="Basescape\SoldierTransformationListState.h">
       <Filter>Basescape</Filter>
-    </ClInclude>
-    <ClInclude Include="Geoscape\AltMonthlyReportState.h">
-      <Filter>Geoscape</Filter>
     </ClInclude>
   </ItemGroup>
   <ItemGroup>
