+++ src/Savegame/SavedGame.cpp
@@ -146,7 +146,7 @@ bool haveReserchVector(const std::vector<const RuleResearch*> &vec,  const std::
 SavedGame::SavedGame() : _difficulty(DIFF_BEGINNER), _end(END_NONE), _ironman(false), _globeLon(0.0),
 						 _globeLat(0.0), _globeZoom(0), _battleGame(0), _debug(false),
 						 _warned(false), _monthsPassed(-1), _selectedBase(0), _autosales(), _disableSoldierEquipment(false), _alienContainmentChecked(false),
-						 _loyalty(0), _lastMonthsLoyalty(0)
+						 _loyalty(0)
 {
 	_time = new GameTime(6, 1, 1, 1999, 12, 0, 0);
 	_alienStrategy = new AlienStrategy();
@@ -430,7 +430,6 @@ void SavedGame::load(const std::string &filename, Mod *mod, Language *lang)
 	_graphCountryToggles = doc["graphCountryToggles"].as<std::string>(_graphCountryToggles);
 	_graphFinanceToggles = doc["graphFinanceToggles"].as<std::string>(_graphFinanceToggles);
 	_loyalty = doc["loyalty"].as<int>(_loyalty);
-	_lastMonthsLoyalty = doc["lastMonthsLoyalty"].as<int>(_lastMonthsLoyalty);
 	_funds = doc["funds"].as< std::vector<int64_t> >(_funds);
 	_maintenance = doc["maintenance"].as< std::vector<int64_t> >(_maintenance);
 	_userNotes = doc["userNotes"].as< std::vector<std::string> >(_userNotes);
@@ -866,7 +865,6 @@ void SavedGame::save(const std::string &filename, Mod *mod) const
 	node["graphFinanceToggles"] = _graphFinanceToggles;
 	node["rng"] = RNG::getSeed();
 	node["loyalty"] = _loyalty;
-	node["lastMonthsLoyalty"] = _lastMonthsLoyalty;
 	node["funds"] = _funds;
 	node["maintenance"] = _maintenance;
 	node["userNotes"] = _userNotes;
